update radial convert

+++ GUI Dia Vision.py
@@ -5,6 +5,7 @@ from PyQt5 import QtGui
 import sys
 import Defenation_for_Countours
 import Detect_countours
+from Radial_Vision import radial_convert
 
 class dlgMain(QDialog):
     def __init__(self):
@@ -12,10 +13,11 @@ class dlgMain(QDialog):
         super().__init__()
 
         self.res = [""]
+        self.var = 0
 
         self.setWindowIcon(QtGui.QIcon("icon.png"))
         self.setWindowTitle("Dia Vision")
-        self.resize(450, 220)
+        self.resize(450, 320)
 
         self.label = QLabel("Введите путь до файла: ", self)
         self.label.setFont(QFont("Arial", 11))
@@ -34,11 +36,28 @@ class dlgMain(QDialog):
         self.editText.setFont(QFont("Arial", 9))
         self.editText.move(240, 80)
 
+        self.vb1 = QPushButton("Radial", self)
+        self.vb1.setFont(QFont("Arial", 11))
+        self.vb1.move(80, 140)
+        self.vb1.clicked.connect(self.kn1)
+
+        self.vb2 = QPushButton("Vert Line", self)
+        self.vb2.setFont(QFont("Arial", 11))
+        self.vb2.move(260, 140)
+        self.vb2.clicked.connect(self.kn2)
+
         self.final_button = QPushButton("Компелировать график", self)
         self.final_button.setFont(QFont("Arial", 11))
-        self.final_button.move(115, 150)
+        self.final_button.move(120, 210)
         self.final_button.clicked.connect(self.final_func)
 
+    def kn1(self):
+        self.var = 1
+        a = QMessageBox.information(self, "info", "Успешно выбрано радиальное конвертирование")
+
+    def kn2(self):
+        self.var = 2
+        a = QMessageBox.information(self, "info", "Успешно выбрано линейное конвертирование")
 
     def vybor(self):
         self.res = QFileDialog.getOpenFileName(self, "open", r"C:\Users", "Graphic png(*.png)")
@@ -50,16 +69,21 @@ class dlgMain(QDialog):
 
     def final_func(self):
         if self.res[0] == '':
-            az = QMessageBox.warning(self, "Warning", "Вы не выбрали файл")
+            az = QMessageBox.warning(self, "Warning", "Вы не выбрали файл ")
         if self.editText.text() == "":
             ax = QMessageBox.warning(self, "Warning", "Вы не дали название ")
+        if self.var == 0:
+            azzzx = QMessageBox.warning(self, "Warning", "Вы не выбрали тип конвертирования")
 
-        result = Detect_countours.coo(len(Defenation_for_Countours.pixel_cont(Defenation_for_Countours.rotata(self.res[0]))), Defenation_for_Countours.rotata(self.res[0]))
+        if self.var == 2:
+            result = Detect_countours.coo(len(Defenation_for_Countours.pixel_cont(Defenation_for_Countours.rotata(self.res[0]))), Defenation_for_Countours.rotata(self.res[0]))
 
-        Detect_countours.cont_to_ex(self.editText.text(), Detect_countours.raspk_coort(result))
+            Detect_countours.cont_to_ex(self.editText.text(), Detect_countours.raspk_coort(result))
 
-        os.remove(Defenation_for_Countours.rotata(self.res[0]))
+            os.remove(Defenation_for_Countours.rotata(self.res[0]))
 
+        elif self.var == 1:
+            radial_convert(self.res[0], self.editText.text())
         azx = QMessageBox.information(self, "Успех", "График успешно скомпелирован")
